fix(DEV-14109): issue and send fixed on the preview page
